+++ plugins/auth-backend-module-github-provider/src/authenticator.ts
@@ -28,7 +28,10 @@ const ACCESS_TOKEN_PREFIX = 'access-token.';
 export const githubAuthenticator = createOAuthAuthenticator({
   defaultProfileTransform:
     PassportOAuthAuthenticatorHelper.defaultProfileTransform,
-  shouldPersistScopes: true,
+  scopes: {
+    persist: true,
+    required: ['read:user'],
+  },
   initialize({ callbackUrl, config }) {
     const clientId = config.getString('clientId');
     const clientSecret = config.getString('clientSecret');