+++ packages/core-compat-api/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/core-compat-api
 
+## 0.2.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.6.6-next.2
+  - @backstage/core-plugin-api@1.9.3-next.0
+  - @backstage/version-bridge@1.0.8
+
 ## 0.2.6-next.1
 
 ### Patch Changes