+++ packages/core-components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/core-components
 
+## 0.14.8-next.2
+
+### Patch Changes
+
+- 59cee81: Use `inherit` variant on OverflowTooltip underlying Typography component.
+- 83c4251: Adds icons to status component
+- Updated dependencies
+  - @backstage/config@1.2.0
+  - @backstage/core-plugin-api@1.9.3-next.0
+  - @backstage/errors@1.2.4
+  - @backstage/theme@0.5.6-next.0
+  - @backstage/version-bridge@1.0.8
+
 ## 0.14.8-next.1
 
 ### Patch Changes