+++ plugins/auth-backend-module-azure-easyauth-provider/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-auth-backend-module-azure-easyauth-provider
 
+## 0.1.2-next.2
+
+### Patch Changes
+
+- d44a20a: Added additional plugin metadata to `package.json`.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/plugin-auth-node@0.4.14-next.3
+  - @backstage/catalog-model@1.5.0
+  - @backstage/errors@1.2.4
+
 ## 0.1.2-next.1
 
 ### Patch Changes