+++ plugins/auth-backend-module-microsoft-provider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-auth-backend-module-microsoft-provider
 
+## 0.1.14-next.2
+
+### Patch Changes
+
+- 8efc6cf: Added support for the new shared `additionalScopes` configuration.
+- d44a20a: Added additional plugin metadata to `package.json`.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/plugin-auth-node@0.4.14-next.3
+
 ## 0.1.14-next.1
 
 ### Patch Changes