+++ plugins/auth-backend-module-oauth2-proxy-provider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-auth-backend-module-oauth2-proxy-provider
 
+## 0.1.12-next.2
+
+### Patch Changes
+
+- d44a20a: Added additional plugin metadata to `package.json`.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/plugin-auth-node@0.4.14-next.3
+  - @backstage/errors@1.2.4
+
 ## 0.1.12-next.1
 
 ### Patch Changes