+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @backstage/plugin-auth-backend
 
+## 0.22.6-next.3
+
+### Patch Changes
+
+- d44a20a: Added additional plugin metadata to `package.json`.
+- 3e1bb15: Updated to use the new `@backstage/plugin-auth-backend-module-onelogin-provider` implementation
+- Updated dependencies
+  - @backstage/plugin-auth-backend-module-onelogin-provider@0.1.0-next.0
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/plugin-auth-node@0.4.14-next.3
+  - @backstage/plugin-auth-backend-module-atlassian-provider@0.2.0-next.2
+  - @backstage/plugin-auth-backend-module-bitbucket-provider@0.1.2-next.2
+  - @backstage/plugin-auth-backend-module-github-provider@0.1.16-next.2
+  - @backstage/plugin-auth-backend-module-gitlab-provider@0.1.16-next.2
+  - @backstage/plugin-auth-backend-module-google-provider@0.1.16-next.2
+  - @backstage/plugin-auth-backend-module-microsoft-provider@0.1.14-next.2
+  - @backstage/plugin-auth-backend-module-oauth2-provider@0.2.0-next.2
+  - @backstage/plugin-auth-backend-module-oidc-provider@0.2.0-next.3
+  - @backstage/plugin-auth-backend-module-okta-provider@0.0.12-next.2
+  - @backstage/plugin-auth-backend-module-cloudflare-access-provider@0.1.2-next.3
+  - @backstage/plugin-auth-backend-module-azure-easyauth-provider@0.1.2-next.2
+  - @backstage/plugin-auth-backend-module-oauth2-proxy-provider@0.1.12-next.2
+  - @backstage/plugin-auth-backend-module-aws-alb-provider@0.1.11-next.3
+  - @backstage/plugin-auth-backend-module-gcp-iap-provider@0.2.14-next.2
+  - @backstage/plugin-catalog-node@1.12.1-next.2
+  - @backstage/backend-common@0.23.0-next.3
+  - @backstage/catalog-client@1.6.5
+  - @backstage/catalog-model@1.5.0
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/types@1.1.1
+
 ## 0.22.6-next.2
 
 ### Patch Changes