+++ plugins/scaffolder-backend-module-rails/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-scaffolder-backend-module-rails
 
+## 0.4.36-next.3
+
+### Patch Changes
+
+- d44a20a: Added additional plugin metadata to `package.json`.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/integration@1.12.0-next.1
+  - @backstage/plugin-scaffolder-node@0.4.5-next.3
+  - @backstage/backend-common@0.23.0-next.3
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/types@1.1.1
+
 ## 0.4.36-next.2
 
 ### Patch Changes