+++ plugins/scaffolder-backend-module-yeoman/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-scaffolder-backend-module-yeoman
 
+## 0.3.2-next.2
+
+### Patch Changes
+
+- d44a20a: Added additional plugin metadata to `package.json`.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.3
+  - @backstage/plugin-scaffolder-node-test-utils@0.1.5-next.3
+  - @backstage/plugin-scaffolder-node@0.4.5-next.3
+  - @backstage/types@1.1.1
+
 ## 0.3.2-next.1
 
 ### Patch Changes